Move top-level drgn.helpers helpers into drgn.helpers.common submodules
This is a better place to put all of these generic helpers now. I think
these are all uncommon enough that it shouldn't be too much trouble to
move them.
