+++ drgn/helpers/linux/slab.py
@@ -19,7 +19,7 @@ import operator
 from typing import Iterator, Optional, Set, Union, overload
 
 from drgn import NULL, FaultError, IntegerLike, Object, Program, Type, cast, sizeof
-from drgn.helpers import escape_ascii_string
+from drgn.helpers.common.format import escape_ascii_string
 from drgn.helpers.linux.cpumask import for_each_online_cpu
 from drgn.helpers.linux.list import list_for_each_entry
 from drgn.helpers.linux.mm import (
