+++ tests/helpers/common/test_format.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from drgn import Program, TypeEnumerator
-from drgn.helpers import decode_enum_type_flags, decode_flags
+from drgn.helpers.common.format import decode_enum_type_flags, decode_flags
 from tests import MOCK_PLATFORM, TestCase
